+++ types/TimeTable.ts
@@ -9,7 +9,6 @@ export interface TimeTableData {
   generatedDate?: string;
 
   dayNames: string[];
-  shortDayNames: string[];
   days: TableLesson[][][];
   hours: Record<number, TableHour>;
 }
